Integrate custom_label.Jenkinsfile into main pipeline
The custom_label.Jenkinsfile is a duplicate of the main
Jenkinsfile with tests against custom agents labels.
This was supposed to be used for the E2E testing.

Instead of duplicating the main Jenkinsfile, it's
better to have the changes for the custom agents labels
integrated into the main Jenkinsfile.

The backwards compatibility is maintained with this
change.

Signed-off-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>